pyproject.toml
[tool.poetry]
name = "pylang"
version = "0.1.0"
description = ""
authors = ["Michael Caley <mjcaley@darkarctic.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
lark-parser = "^0.6.5"
llvmlite = "^0.26.0"

[tool.poetry.dev-dependencies]
pytest = "^4.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
